Add wmerrors.error_script_file
Add the php.ini setting "wmerrors.error_script_file", which specifies a
PHP script to be executed when a fatal error is encountered. This allows
error logging and display to be customised in userspace.

I'm not completely sure it won't crash. The theory is that it's better
to have backtraces most of the time and segfaults occasionally than no
backtraces at all.

Also:
* Remove a few TSRMLS_CC, these are unnecessary in PHP 7
* Update a couple of comments
* Remove error.html, a very old copy of the WMF error page

Bug: T187147
Change-Id: Ia679692d41d4c83c556ce69b8eeb92a0bca34a3b