Bump version to 2.0.0
When I packaged the initial git snapshot for Debian in 2019, I set
a version number of 2.0.0. I assume this was because the changes for
7.2 support were a major change from the previous PHP5-era code.

Anyways, the easiest way to fix this is to set the version to 2.0.0
now. Given that we've never properly tagged a version, it seems like
a good idea to consider the addition of PHP 8 support a major change.

Bug: T293932
Change-Id: I9e7e10416f826a1aa7e3d202c6a58ef3cfebb4af